+++ NEWS.md
@@ -148,6 +148,9 @@
   
 * `min(.,na.rm = TRUE)` works with `Date`s built on numeric vectors (#755)
 
+* Internal implementations of `dense_rank`, `min_rank`, `cume_dist` and 
+  `percent_rank` deal correctly with NA (#774)
+
 # dplyr 0.3.0.1
 
 * Fixed problem with test script on Windows.